requirements.txt
crewai[tools]
pandas
matplotlib
seaborn
python-dotenv
colorama
reportlab
pydantic
numpy
langchain-google-genai
scikit-learn
requests 